api/pyproject.toml
[project]
name = "authapp"
version = "0.1.0"
description = ""

authors = [
    { name = "willian", email = "willian.wyann@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.13"

dependencies = [
    "flask (>=3.1.0,<4.0.0)",
    # flask restful 风格切换
    "flask-restful (>=0.3.10,<0.4.0)",
    # 数据验证框架
    "pydantic (>=2.10.6,<3.0.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    # 数据库连接
    "flask-sqlalchemy (>=3.1.1,<4.0.0)",
    "pymysql (>=1.1.1,<2.0.0)",
    # 数据库表迁移
    "flask-migrate (>=4.1.0,<5.0.0)",
    "flask-sqlacodegen (>=2.0.0,<3.0.0)",
    "cryptography (>=44.0.2,<45.0.0)",
    # 登录认证
    "flask-login (>=0.6.3,<0.7.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    # redis
    "redis (>=5.2.1,<6.0.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"